// src/app/models/enum.ts
export enum Gender {
  MALE = 'MALE',
  FEMALE = 'FEMALE',
  NO_REFERENCE = 'NO_REFERENCE',
}

export enum ReferralCode {
  TEAM_A = 'TEAM_A',
  TEAM_B = 'TEAM_B',
  UNKNOWN = 'UNKNOWN'
}

export enum UserType {

  MEMBER = 'MEMBER',
  TEAM_LEADER = 'TEAM_LEADER',
  OWNER = 'OWNER',
  EMPLOYEE = 'EMPLOYEE',
  ADMIN = 'ADMIN',

}

export enum UserStatus {
  PENDING = 'PENDING',
  ACTIVE = 'ACTIVE',
  DISABLED = 'DISABLED'
};